server: Implement the --emit-server-status=<DEST> server option (#2463)
The new `--emit-server-status=<DEST>` option replaces
`--echo-runtime-info`.  Unlike the latter, `--emit-server-status`
accepts an explicit file path or a file descriptor to emit server
information to.  This lets server `stdout` be undisturbed, and adds
a way for future additions such as reporting server status via a
socket, which might be useful for a service supervisor.

This also does a bunch of cleanups in the `server_ops` tests.