Implement pool "Garbage Collection" (discard unused conns)
This commit makes the per-block connection queue *a stack*, with each
connection paired with a monotonic timestamp indicating when it was
pushed into the stack. GC is implemented using the same `try_steal`
logic to pull connections from the bottom of the stack until a preset
timestamp limit is met. GC is now running every 2 minutes after the
first idle connection is freed into the pool, and pauses if no idle
connection is freed in the last 2 GC cycles. So theoretically an idle
connection will be discarded between 2-4 minutes after it's returned to
the pool and not otherwise stolen or rebalanced.