Remove all pgsql backend renaming code
Since everything in the backend uses uuids for names and all
user-defined objects live in the same pgsql schema, we don't need any
of our backend renaming code, so rip it all out.

If something changes in the future and we need to generate a pgsql
rename, we can pull the old dbops code out of git.

Fixes #2406.