Pool optimization (#2480)
* Single block: don't create connection if more than 1 available

This design balances between connection overhead and total number of
open connections when we're managing connections to only one DB.
If possible, we always try to keep 1 spare connections - no more,
no less - in the pool for the next acquire. So we don't open too many
unused connections (will be GC-ed anyway), or hold the next acquire
waiting to open new connections.

Also added base test in the simulation testing.

* Don't rebalance if max_capacity is not reached.

If we still have space for more connections, don't actively
rebalance the pool just yet - rebalance will kick in when the max
capacity is hit; or we'll depend on the garbage collection to
shrink the over-quota blocks.

* Implement pool "Garbage Collection" (discard unused conns)

This commit makes the per-block connection queue *a stack*, with each
connection paired with a monotonic timestamp indicating when it was
pushed into the stack. GC is implemented using the same `try_steal`
logic to pull connections from the bottom of the stack until a preset
timestamp limit is met. GC is now running every 2 minutes after the
first idle connection is freed into the pool, and pauses if no idle
connection is freed in the last 2 GC cycles. So theoretically an idle
connection will be discarded between 2-4 minutes after it's returned to
the pool and not otherwise stolen or rebalanced.

* pool: Add comments and trivial fix

Also added a simplified `_steal_conn()` to the `NaivePool` so that it
doesn't hang if a new block is requested while there's no connection
returned in old blocks.

* Prune unused connections when creating from template database

This is similar to what we do before dropping a database. Postgres would
raise the (same) error if we don't prune the unused connections. This is
only a problem now because of the recent changes we made to the pool.