Support filling a newly created required field
This involves adding syntactic support for using SET REQUIRED USING in
CREATE PROPERTY/LINK and making schema diffing able to understand it.

Partially fixes #2486. Unfortunately it doesn't actually fix the issue
that the bug was filed originally as, which is prompting to populate a
required field that is being added by a new parent type.