Fix migration prompting while adding required pointers via parent types
The key idea is that we force a `CREATE PROPERTY ... { SET REQUIRED  ... }`
to be generated when appropriate even if we would normally elide it.

Forcing this to happen exposed a funny corner case in which the
diffing mechanism sometimes generates DELETE/CREATE pairs for the
__type__ link when a type is renamed. This might be worth dealing with
more robustly, but I worked around this for now by having the compiler
regenerate the proposed plan after applying DDL that does a rename.
This turns out to have been a good idea in general, since it fixed
issues where we would generate prompts that referred to old names
instead of new ones.

Fixes #2486.