Prevent illegal UPDATEs to covariantly overloaded link types
Insert dynamic type checks in situations where it can occur.

One slightly fiddly part is making sure that we can extract the
__type__ from the places we need it (since doing this check can
require __type__ in places where it otherwise wouldn't appear anywhere
in the AST). This is solved with a bit of a hack: UpdateStmt is
augmented with a PointerRef to BaseObject.__type__, from which we can
extract whatever we need.

Also fixes creation of covariantly overloaded multi links.

Fixes #2324.