Support switching to binary protocol with HTTP upgrade
The server now follows HTTP/1.1 upgrade protocol to promote a single
frontend connection from HTTP-based protocol to EdgeDB binary protocol.
The protocol is named `edgedb-binary` and it has to be exactly the same
in the Upgrade header, e.g. `Upgrade: edgedb-binary`.

Previous mechanism to detect if a new connection is HTTP or binary by
looking at the first byte still remains, so old clients are still
supported. But new clients are suggested to use HTTP upgrade on
connect in order to switch to the binary protocol (or remain in HTTP).