Drop the reverse_map_path_id in _maybe_get_path_rvar
The map reversing there I think doesn't make any sense; if it finds
something, there's no guarentee that the thing it finds will have the
right outputs, since /it/ doesn't know about the mapping.

The one impact it had on our test suite was that it was *preventing*
us from finding something in an UNLESS CONFLICT case where we do an
INSERT in the ELSE, but that was just an accident.