Prevent endpoint property commands from separating from parent link
The `@source` and `@target` properties are special and must be present
for every link through its lifetime, so make sure deletion and creation
commands do not get separated from the parent link command.  This is
achieved by generalizing the static `StronglyReferencedObjectCommand`
into a `is_strong_ref` field on `ReferencedObjectCommand` commands.