Perform target type cleanup on @target property deletion instead of link deletion
Currently, generated types, such as compound types get cleaned up in the
`DeleteLink` handler, but the link object is not the only reference to
the target type: there is also the special `@target` property, which
will erroneously keep the target type alive, because `_has_outside_refs`
does not account for these when checking target deletion in the context
of link deletion directly.  Fix this by triggering target type deletion
on `@target` removal instead.