schema: Untangle class-specific expr ref finalization code
Currently, expr ref fixup setup and finalization is done in
`ObjectCommand` in a large switch that imports half of the schema
modules.  There is no particular reason why this needs to be this way,
so move fixup specializations into appropriate delta commands as the new
`canonicalize_alter_from_external_ref()` and `get_dummy_expr_field_value()`
methods.  This should reduce the amount of surprising action at a distance
and allow for better code locality and reuse.