Materialize volatile (and some other) sets in binding constructs
Materializing a set of values consists of:
 * Serializing each of the values in native mode, but with computed
   shape fields included.
 * Then, if the value is possibly multi, aggregating it into an array.

We consider materializing a set of values in two of our core binding
constructs:
 * WITH statements
 * ad-hoc computed pointers.
 * We don't materialize the argument to FOR, because I thought that
   FOR's iteration makes it unnecessary. I found some bugs involving
   FOR, though, so maybe that is wrong.

We materialize a set in the following cases:
 * If it is volatile, or uses a bound variable that is volatile
 * If it appears in a computable and refers to a bound variable.
   This is somewhat overly broad, but is intended to allow us to
   capture FOR variables inside a shape, and similar patterns.

The materialized sets are computed at the enclosing statement:
at the WITH statement, or at the shape subject expression.

We detect when a materialized set is referenced by its type/ptrcls.

To support this, queries on the pgast side gain some new fields for
tracking materialized (or packed) data. packed_path_rvar_map tracks
what rvars contain materialized data while packed_path_outputs
contains the actual output vars.

The most involved part of this is unpack_rvar, which extracts a set
from a packed rvar. It inspects the type of the set, generates SQL to
unpack it, and populate the current statement with the appropriate
rvars for accessing it and its fields. Materialized links (to
materialized shapes) and multi properties of shapes are placed in
packed rvars of their own.

Generating path_ids for tuple paths and shape paths on the fly, in
pgsql, is at least a little hair-raising. (Some dubious machinery
needed to be added to typeutils for the latter.)

In order to ensure that shapes (and tuples that contain shapes) can be
correctly output or correctly *rematerialized*, we need to reconstruct
a serialized version when we unpack. Doing this for tuples means we
need to do it for any shapes they contain, which means that unpack
needs to be driven by a recursive traversal of the type.
This also drives a reconstruction of the *value* of tuples, so that
a tuple containing objects can safely compared for equality and the
like, without inadvertantly comparing materialized shape data.

Known issues:
 * We are a little over-eager to materialize. There are many cases
   where we only use a value once where we will materialize.
   (But also cases where we only use values once where we need to.)
 * We materialize AT_MOST_ONE sets as if they were MANY, serializing
   them into an array. This was an implementation expedient; it
   shouldn't be too hard to use NULL instead.

 * Materializing arrays that contain shapes is pretty busted. The
   tricky part is that unpacking rvars with shapes works by populating
   the current rel with rvars for each of the shape fields. This is
   a problem with arrays, because we can't get at those until after
   some elimination form has been done on the array, and there are
   several of those we'll need to deal with.

 * As I was doing final cleanups on this PR, I found a big cluster of
   bugs involving FOR/tricky uses of WITH (
   test_edgeql_volatility_for_like_hard_*,
   test_edgeql_volatility_for_hard_*).
   I am fairly concerned about these, but didn't want to delay
   getting a PR up for review any longer.

Testing this was a little fiddly, since assert_query_result is pretty
useless when the output is volatile. It was important to test in both
json and native mode (since I had regularly bugs that broke only one
of them!), so I ended up building some of my own infrastructure for that.

This fixes some existing tests (including test_edgeql_scope_tuple_14),
though not as many as I hoped. (I was hoping to get some of the FOR
tests, which I'll hopefully take more of a look at soon.)

Closes #2421.

Co-authored-by: Elvis Pranskevichus <elvis@edgedb.com>