Adjust the precedence of DETACHED to match EXISTS/DISTINCT (#2638)
Currently the situation is that `SELECT DETACHED User { x := User.name }`
parses as `SELECT (DETACHED User) { x := User.name }`,
but the reference to `User` in the shape is still expected to refer
to the same `User` being selected over. (See
`test_edgeql_scope_detached_{07,08}`)

I claim that this makes no sense: `User` is DETACHED, and nothing
outside that detached subexpression ought to be able to correlate with
it.
If we want references inside the body of the shape to correlate, then,
we need to raise the precdence of `DISTINCT` so that the above example
parses as `SELECT DETACHED (User { x := User.name })`.

This also helps solve some other issues with the current implementation:
 * Currently the example query I gave doesn't quite work: it infers
   `x` as multi, and returns the results as sets
 * `SELECT (User.name, { x := User.name })`, which is sugar for
   `SELECT (User.name, (DETACHED VirtualObject) { x := User.name })`,
   ISEs at SQL execution, because the cardinality of `x` is inferred as
   single but `x` is compiled with `User` detached.

These issues are caused by a mismatch between `process_view`, which does
not adopt the namespace of its source, and `_get_computable_ctx`, which
does. In this change, we wrap a DISTINCT expression in another set, to
prevent its freshly generated namespace from leaking out.