Get rid of the remaining @overload duplication in schema (#2670)
For all the overloaded getFOO methods, introduce an abstract _getFOO
method to be implemented by children and dispatch to it from the
abstract base. This lets us ditch listing the overloads in the children.

`get` already got this treatment, but the function was named
`get_generic`. Change `get_generic` to `_get` and the existing `_get`
in `FlatSchema` to `_search_with_getter`.