Implement std::assert_single()
The new `std::assert_single(input: SET OF anytype) -> anytype` function
checks the cardinality of the input set, and if it is greater than 1,
raises a `CardinalityViolationError`.  This runtime cardinality
assertion is intended to be used in place of the current `LIMIT 1` hack
to safely place a singleton cardinality invariant on an expression where
static inference cannot help.

Fixes: #2402