Rename CONFIGURE SYSTEM to CONFIGURE INSTANCE
CONFIGURE SYSTEM proved to be confusing because we have no concept of
"system" anywhere but in EdgeQL, currently. We do have a concept of
"instances" in our tooling, so we're replacing SYSTEM with INSTANCE to
make things consistent.

The old CONFIGURE SYSTEM command is kept working, although we will start
issuing deprecation warnings for it soon.