Fix some materialization related issues
The main thing here is to rework the selection of which materialized
set is compiled to ensure that we always compile one with an
appropriate rptr source, if it is a computable.

Additionally we need to store materialized computables on a non-SELECT
shape with the shape statement, not with whatever edgeql statement is
surrounding it, which requires shifting how things are tracked.

Modify the test_volatility tests to also insert typenames, and fix
some issues that shook out.

As part of dealing with the changes to non-SELECT shape compilation, I
needed to make the visibility checks in cardinality inference and
subquery compilation be done from the parent *branch* instead of the
parent *fence*. I am not super confident about this change.