Prohibit navigational nesting in UPDATE (#2728)
Currently, we allow the following `UPDATE` syntax:

```
UPDATE Foo
FILTER .foo_name = 'First'
SET {
   bar: {
       spam := true
   } FILTER .bar_name = 'Blah'
}
```

The intuitive interpretation is that it should update an object of name
`Blah` specifically linked to a `Foo` object of name `First`.  However,
currently this is what the statement gets transformed into:

```
UPDATE Foo
FILTER .foo_name = 'First'
SET {
   bar := (SELECT .bar {
       spam := true
   } FILTER .bar_name = 'Blah')
}
```

It's the standard `SELECT` transform, but it's quite wrong for `UPDATE`,
since instead of updating a single linked object, it actually _replaces_
the entire linked set `bar`.

We intend to support the nested `UPDATE` case properly soon, but for now
we should just ban the syntax.

While at it, fix a bug whereby UPDATEs that reassign a link to an
overlapping set don't get the link properties erased properly.