Fix applying changes to listen_addresses
Now we'll have one TCP server for each `listen_addresses`, and updating
`listen_addresses` will only start the new servers and stop the servers
that is not in the update.

At the same time, we are broadcasting new `listen_addrs` in the
`--emit-server-status` file or socket, because `--port=auto` may lead to
different ports bound on different hosts. The admin UNIX socket will
not be created until there is a single TCP port.

Because a single `listen_address` may be resolved into multiple IPs and
this PR left the resolution work to asyncio `start_server()`, there is a
case when the `listen_addresses` is updated to a different name but
resolves into the same IP, the issue described in #2696 would still
happen.

Fixes #2696