Adopt "computed" link/property instead of "computable".
The "computed" link or property is a little more straightforward name
for the comcept. The link or property are indeed computed by an
expression as opposed to being statically defined. It also appears that
this term is used to denote similar concepts elsewhere and thus will be
easier to search for and understand.

Update the docs, error messages and tests.