Add a command-line switch to allow insecure HTTP client connections.
The plan is to pass it automatically to instances started via the
`edgedb instance` / `edgedb project` tools.  The motivation for
that is to enable local development.  Without unsecure HTTP, users
will not be able to browse the integrated GraphiQL or send
GraphQL/EdgeQL HTTP requests to http://localhost.

The new argument is called '--allow-insecure-http-clients'. While
there, rename the earlier introduced '--allow-cleartext-connections'
to '--allow-insecure-binary-clients' for consistency.