Be smarter about how and when we materialize
Instead of always materializing whenever we see a binding in a shape
element, we track the *reasons* we think we might need to
materialize. While finalizing materialization decisions, we look at
the bindings used, and only materialize when a *visible* binding is
referenced.

This means that doing WITH/FOR bindings internally is fine, but
referring to externally bound things needs materialization.

My main motivation for doing this is that the approach I am planning
to take to fix the nesting tests introduced in #2774 involves
materializing those cases, but we obviously can't just materialize
everything that references some unrelated object, so we need machinery
to base it on scope visibility.

A basic example of where this helps is:
```
WITH U := User { asdf := (WITH D := .deck, SELECT D.name) },
SELECT U.asdf;
```
Previously we would do two `array_agg`s, for `asdf` and `U`, and now
we do none.

Another important side effect of this is a WITH bound set of a shape
that has a pointer materialized for binding reasons does not need to
be materialized itself!
(This required having `maybe_materialize` search through the view
pointer hierarchy to find relevant materialized pointers through
multiple views.)

Some queries that this helps with:
```
WITH X := (FOR x in {"!","?"} UNION (User { z := .deck.name ++ x}))
SELECT X {name, z} FILTER .name = "Alice";
```
The `z` multi property is materialized into an array, but `X` does not
have to be.

We can even do:
```
WITH X := (FOR x in {"!","?"} UNION (User { z := .deck.name ++ x}))
SELECT ((SELECT X.z), (SELECT X.z));
```
The whole computation gets duplicated (as is traditional in edgedb)
and the `z` property gets materialized in both.

Unfortunately right now this one still materializes:
```
WITH U := (FOR x in {"!","?"} UNION (User { z := .deck.name ++ x})),
     X := (SELECT U FILTER .name = 'Alice'),
SELECT ((SELECT X.z), (SELECT X.z));
```