Fix a couple of newly reported mypy errors
For bizarre and unknown reasons, mypy suddently became unhappy about two
code locations:

    AssertionError: mypy validation failed:
    edb/schema/schema.py:817:5: error: Missing return statement
    edb/schema/schema.py:1925:9: error: Return value expected

The second is a bare `return`, so fair enough, unclear why it was OK
before, while the first looks like a mypy bug to me.  A `return`
statement in the `with` body should be perfectly valid.  I suspect there
is some weird interaction with new annotations in `immutables`.