Turn assert_single into a no-op when its argument is provably a singleton
This changes the implementation of `assert_single()` to turn it into a
no-op when its argument has been statically proven to be a singleton.
This is beneficial for language query builders, since in lieu of full
cardinality inference on the client side, they'd have to rely heavily on
cardinality assertions, so we want them as cheap as possible.