Change how JSON scalars are reflected into GraphQL.
`std::json` is now reflected into a custom JSON scalar type in Graphql.
This is an escape hatch so the output is rendered as seamless JSON (not
as a string encoding JSON). The input values can only be supplied as a
variable at the moment (because of issues of parsing and validating
GraphQL types otherwise).