Consolidate tool configuration into pyproject.toml
This moves `mypy.ini` and `.coveragerc` into `pyproject.toml`.  Aside
from fewer configuration files in the root directory, this is a massive
improvement for mypy, as we don't have to deal with its torturous module
selection syntax anymore.  flake8 is the only straddler here, and it
seems its maintainers are aggressively against supporting
`pyproject.toml` for configuration for now [1]

[1] https://github.com/PyCQA/flake8/issues/234