Use asyncio subprocess API in PostgreSQL cluster management
This switches all uses of `subprocess` in `pgcluster` to
`asyncio.create_subprocess_exec` and sets up consistent capturing and
logging of subprocess standard streams.

Some slight massaging of log formatting is also here: timestamp is now
formatted as proper ISO date, and logger names are a darker shade of
grey.

Fixes: #2819