Fix some backlink cases (#2840)
This was prompted by an issue that came up with backlinks in
introspection queries, which was caused by logic errors in the code
that arranges to output the source path of an inline backlink in the
case where the typeref rvar is a subquery pointing to a CTE. While
investigating I also found a handful of issues involving
union/intersection interactions.

To fix this, I deleted all of that code and arranged things to use
`get_path_output` to find the relevant columns. This unfortunately
required adding a `outbound_ptr` to `BasePointerRef` to allow us to
invert a backlink. I think that it might be better if PointerRef didn't
have a direction on it, and that direction was always tracked externally
(which it is also, with Pointer).

Fixes #2599.