Fix cardinality inference in "optionality-preserving" functions (#2844)
A certain number of standard aggregate functions are guaranteed to
preserve the lower cardinality bound of their argument.  These functions
are `std::assert_single`, `std::enumerate`, `std::min`, and `std::max`.
Unfortunately, there is currently no way to signal that lower
cardinality is preserved, and so we introduce the new std-only
`preserves_cardinality` field, which is also exposed in the public
introspection schema for the benefit of query builders and such.

While doing this, I realised that we also mishandle functions with
non-OPTIONAL arguments that return `OPTIONAL` or `SET OF`, so I fixed
inference of those too.