Drop FINAL from the syntax
The `final` attribute has some internal uses, but they don't exactly
correspond to what FINAL ought to mean in the language. It's
undocumented, so just yank it.

The one snag is that we actually currently when describing DDL will
*always* describe a scalar type as being FINAL, so dumps will contain
`CREATE FINAL SCALAR TYPE`. This means that we need to continue accepting
that particular syntax, but disallow it based on our compat version.
Ugly, but oh well. If we later add a proper FINAL, then we'll continue
ignoring FINAL when the compat ver is old enough.

Fixes #2607.