Add a (failing) test for generic overloads in SDL
The current `Function.backend_name` logic is trying is magic too early,
and so if one attempted to make generic overloads in SDL, the generated
functions would not share a `backend_name` and dispatch from a generic
function body would call the wrong function.

(To be fair we currently disallow generic functions in user code, so I
was only able trip over this in test mode, along with a general
mishandling of pseudo-types in SDL).