Fix children detection in ON CONFLICT compilation
A bug there caused us to *always* use conflict CTEs for UNLESS
CONFLICT, which is probably somewhat inefficient and was unintended.

This papers over #2869 (and solves the immediate problem of the
edgedb-python test breakage), because the switch to using conflict
CTEs caused some tests of concurrent UNLESS CONFLICT transactions to
fail with ConstraintViolationErrors.

There are plenty of other situations where we *need* conflict CTEs,
though, so I think this one is gonna give us some heartburn.