Always allow FINAL for enum CREATE SCALAR TYPEs
It's written into migrations and dumps for enums, so we need to
support it somehow.

There is still a potential backward compatability hazard for when we
actually introduce FINAL, which is if anybody actually has a non-enum
scalar that extends an enum. That doesn't really accomplish anything
so maybe we don't care.