Handle exclusive violations caused by UPDATEs to related types
This expands the infrastructure from #2881 to also cover
UPDATEs. UPDATEs are trickier than inserts for a number of reasons:
 1. Since UPDATE can update child objects as well, it can fail because
    of constraints only present on child types (or even on other
    ancestors of child types).
 2. Because of this, UPDATEs can conflict with themselves, so
    all UPDATEs of types with children and constraints need this
    treatment.
 3. The conflicts can involve data that doesn't appear in the query
    itself.

Handling all the subtypes requires looping over all subtypes on both
sides of a potential conflict, as well as populating overlays for all
subtypes after an update. (The latter turned out to be necessary for
other reasons too, though.)

Handling the data that doesn't appear in the query itself required
arranging to use the CTEs as the source of truth for what is being
updated rather than pulling it from the query.

I also did some optimization of how we look at conflicting data, by
stripping out filters and the base relation from the overlays. This
was necessary to avoid a 25s test suite perf regression on my machine.

Fixes #2845.