Fix handling of function overloads on object type parameters (#2889)
Currently, an attempt to declare a function overload with a
difference in an object type parameter crashes with an ISE:

    Error: InternalServerError: function "..." already exists with same argument types

This is because objects are passed to functions by identity and Postgres
is unable to distinguish between the implementations, since all would
have the same `uuid`-typed parameter.

A straightforward solution for this would involve adding a dummy
argument of the composite type corresponding to the object type
relation.  This will only work for compile-time dispatch, i.e.
`SELECT foo(Obj)` would always call the `foo(Obj)` variant, even if more
specific `foo()` implementations exist for subtypes of `Obj`.  This
behavior is very unintuitive and so we must do runtime function dispatch.
Unfortunately, there is no way to do such dispatch in Postgres without
resorting to `EXECUTE` in PL/pgSQL, which would be very slow, so the
solution here is to put all overloaded implementations into a single
function body and switch based on the `__type__` value passed as an
additional shadow argument for the object parameter.

This approach currently places a number of restrictions on overloads of
object type functions: 1) there must be no difference in any but one
parameter type, i.e. this is single dispatch, and 2) the names of
all parameters must match, because all implementations share a single
function definition.  The second restriction is an implementation
caveat, as we can easily normalize the names of positional arguments,
but I decided to keep that out of scope of this patch.