Better error messages on schema type mismatches (#2896)
Make `Schema.get` produce its own error on a type mismatch instead of
using `get_by_id`, since it has access to the name and so can produce
a better message. Also pass in the sourcectx when resolving names as
part of shell construction, so that the errors will have source
contexts.

Fixes #1016.

In declarative, replace some type asserts with type checks and error messages,
and add some infrastructure for producing the error messages.

Fixes #2556.