Enable compile-time multiplicity enforcement for computed links
A computed link should be indistinguishable from a regular link in
queries.  This means that just like regular links, a computed link
_must_ resolve to a distinct set.  After this, it is a compile-time
error to use any expression that could not be proved to have
multiplicity of at most 1.  Queries where static inference is
impossible should use `assert_distinct` or `DISTINCT` depending on
whether duplicates are expected.

Several `for` tests regressed, but the regression does not seem related
to the change here and simply exposed a preexisting bug.

Fixes: #117