Get rid of setuptools-scm
We aren't actually benefiting from `setuptools-scm` anymore because we
call `git` and format our version manually anyway, and avoiding the
`setuptools-scm` dependency removes some of it's transivite deps from
the runtime list.

This also moves `edb.server.buildmeta` to `edb.buildmeta`, because it
represents the entire distribution not just the server part and reduces
the possibility of import issues at `setup.py` time.