Get rid of FOR scope hacks (#2915)
The "iterator_target" metadata together with `iterator_ctx` in context
are remnants of the now removed iterator hoisting.  They're actually
causing incorrect scope such as in the newly added
`test_edgeql_for_in_computable_11`.  This removal umasked a failure in
`test_edgeql_for_in_computable_02`, which is due to issues in how
materialization reasons are tracked across recompilation of shapes.  The
bug is unrelated as is demonstrated by
`test_edgeql_for_in_computable_02b`, which currently fails with or
without the changes in this patch.