Cleanup pgsql/delta delegation of `apply()`
Currently, the base `MetaCommand` class in `pgsql/delta`, which stand
for an adapter for `delta.Command` doesn't actually call
`delta.Command.apply()`, which forces all its subclasses to implement
explicit calling of both `apply()` implementations.  This was
historically needed because of the lack of granularity in delta command
processing flow, however that is no longer an issue, and with little
tweaks it's possible to actually establish normal cooperative delegation
via `super()`.  One minor hitch here is `AlterScalarType`, which has to
deal with a bunch of PostgreSQL DDL restrictions on domains and enums,
that's worked around with a hack for now.

This also makes a bunch of intermediate classes in `pgsql/delta`
unnecessary, I'll clean them up in a subsequent commit.