Add a server option to specify the default authentication method (#2936)
The current default authentication method is hardcoded to `SCRAM` and if
one wants a different default, something like this is currently
required:

    CONFIGURE INSTANCE INSERT Auth {
        priority := 0,
        method := (INSERT Trust),
    };

This is problematic because this configuration is persistent, so it's
harder to start an insecure server temporarily.

This patch adds the new `--default-auth-method` argument as well as its
companion `EDGEDB_SERVER_DEFAULT_AUTH_METHOD` environment variable to
set the default authentication method.  `SCRAM` is still the default,
but devmode servers (`edb cli`) now set to `--default-auth-method=Trust`
as before.