schema: Explicitly segregate inherent and propagation subcommands
Currently, with the exception of prerequisite commands, all subcommands
of a given delta command are stored in a single list, mixing together
inherent subcommands (e.g. creation of standard link properties on
links) and commands resulting in command propagation down the
inheritance tree.  This mixing makes it hard to separate the two stages
of command processing when there is a need to insert extra logic between
the inherent and the propagated parts.  Fortunately, it's quite easy to
separate them by introducing another list, which is what I've done here.

This also makes it explicit that various `_propagate_ref_op`
implementations are not supposed to mutate the schema.