Fix OPTIONAL interaction with shape subqueries
This requires adding an extra null test when compiling shapes, to
prevent emitting shapes containing null for each field, so I also
slightly improved the "nullable" logic to avoid inserting that test
pointlessly all the time.

Adding this check also meant that CONFIGURE INSERT needed to be
modified to not claim that NULL was its id, so I updated it to use a
real UUID, which also meant needing to tweak the overlay processing
to not use RelRangeVar directly.